+++ BackpackAndBeyondBookKeeping/Expense.h
@@ -11,7 +11,7 @@ private:
 public:
 #pragma region Big5
 	//constructor
-	Expense(int cycle = 0, string category = "", string store = "", int numItems = 0, string name = "", int month = 0, int day = 0, int year = 0, double amount = 0.00) 
+	Expense(int cycle = 0, string category = "", string store = "", int numItems = 0, string name = "", int month = 0, int day = 0, int year = 0, int amount = 0) 
 		:Transaction(cycle, category, name, month, day, year, amount)
 	{
 		mStore = store;
